refactor(auth): share ability rule flattening between builders
